Entered Roland Patches -AM First: System.out.println (Math.PI); --> throws Feature not implemented exception. static double d = Math.PI --> as well So I analyzed the error and want to provide a fix for it. (add modifiers to variable declaration)
Second, MCodeUtilities.resolveType resolves wrong if you use a custom class
with names C,... and so on. Try to create a class C and use it in the main
method. I fixed this in

EvaluationVisitor.visit(VariableDeclaration node), and changed
MCodeUtilitiies.resolveType not to use the "C" comparison and so one.

But I am not sure if there are any other implications for it. I saw that you
use this mechanism quite often. BTW, the bug is due to the use of the non
unique string representation in java.lang.Class.getName(). (e.g. Char [] c
resolves to [C, as well as  C [] c -> [C). Perhaps you should think about
relying on unique information, e.g. serialization data.

Third, I fixed an issue regarding final variables in static declarations
(depends on the 1.st bug). I had to pass the modifiers into the mcode and
handle it during variable declaration. E.g. if you have the statement static
final double d = Math.PI. (it did'nt show the value of d before entering the
main method).

Fourth, I fixed a bug when you have an array, e.g. char [] c = ..., and then
says S.o.p (c): there was a typecast error